added error checking to all components (I think) as well as started cleaning up files that are no longer needed
