+++ src/components/Projects.tsx
@@ -1,63 +1,50 @@
-import { useState, useEffect } from 'react'
-import { projectsLink } from '../toolbelt/api'
-import LoadingSpinner from './LoadingSpinner'
+import { useState, useEffect } from "react";
+import { projectsLink } from "../toolbelt/api";
+import LoadingSpinner from "./LoadingSpinner";
+import Accordion2, { Project } from "./Accordion";
 
 function Projects() {
-
-    const [restData, setData] = useState(null)
-    const [isLoaded, setLoadStatus] = useState(false)
+    const [restData, setData] = useState<Project[] | null>(null);
+    const [isLoaded, setLoadStatus] = useState(false);
 
     useEffect(() => {
         const fetchData = async () => {
-            const response = await fetch(projectsLink)
-            if ( response.ok ) {
-                const data = await response.json()
-                setData(data)
-                setLoadStatus(true)
+            const response = await fetch(projectsLink);
+            if (response.ok) {
+                const data = await response.json();
+                setData(data);
+                setLoadStatus(true);
             } else {
-                setLoadStatus(false)
+                setLoadStatus(true); // setting load state to true to avoid infinite spinner (although if they're seeing this then restData has already loaded on main check)
+                setData([]); // setting restData state to an empty array to trigger the "no projects" message incase projects come back empty
             }
-        }
-        fetchData()
-    }, [projectsLink])
+        };
+        fetchData();
+    }, [projectsLink]);
+
+    console.log({restData});
     
-console.log(restData)
-return (
-    <>
-        {isLoaded ? 
-            <>
-                <h3>Featured Work:</h3>
-
-
-                {restData.map(project =>
-                        <article key={project.id} className='project-single'>
-                            <h3>{project.acf.project_title}
-                            </h3>
-                            <img className="project-img" src={project.acf.project_featured_image} alt={project.acf.project_title} />
-                            <p>{project.acf.project_overview}</p>
-                                {project.tools.map((tool, index) =>
-                                <div key={index} className="tools-used">
-                                    <ul className="tools-used">
-                                        <li className="flex justify-center items-center border border-solid rounded-bl-lg rounded-tr-lg w-40 h-10 m-2 single-tool">
-                                            {tool[0]}
-                                            <img className="pl-2 max-w-[36px] h-auto w-auto tool-icon" src={tool[1]} alt={`${tool[0]} icon`} />
-                                        </li>
-                                    </ul>
-                                </div>
-                                )}
-
-
-
-                        </article>
-                        ) } </>
-                     : 
-                        <LoadingSpinner />
-            }
-                    
-    </>
-
-
-        )
+    return (
+        <>
+            {isLoaded ? (
+                restData && restData.length > 0 ? (
+                    <section id="work-section" className="work-section">
+                        <h2 className="text-lg lg:text-3xl font-medium mb-4">Featured Work:</h2>
+                        <Accordion2 projects={restData} />
+                    </section>
+                ) : (
+                    <section id="work-section" className="work-section">
+                        <h2 className="text-lg lg:text-3xl font-medium mb-4">Featured Work:</h2>
+                        <p>No projects to show.</p>
+                    </section>
+                )
+            ) : (
+                <div className="h-screen flex items-center justify-center">
+                    <LoadingSpinner />
+                </div>
+            )}
+        </>
+    );
 }
 
-export default Projects
+export default Projects;
