+++ src/components/RightSide.tsx
@@ -1,6 +1,6 @@
 import About from "./About"
-import Projects3 from "./Projects copy 2"
-import ToolsCarousel from "./Tools copy"
+import Projects from "./Projects"
+import ToolsCarousel from "./Tools"
 import FooterMobi from "./FooterMobi"
 import BottomNavMobi from "./BottomNavMobi"
 
@@ -11,7 +11,7 @@ function RightSide() {
     id="right no-scroll-bar fade-in-slide-up"
     className="overflow-hidden lg:ml-auto lg:w-1/2 lg:border-l-4 lg:border-current lg:px-16 xl:px-32">
 
-        <Projects3 />
+        <Projects />
         <About />
         <ToolsCarousel />
         <FooterMobi />
